microservices/payment-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
stripe==7.11.0
paypalrestsdk==1.13.1
redis==5.0.1
python-dotenv==1.0.0
pymysql==1.1.0
sqlalchemy==2.0.23